requirements.txt
addict
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
torch
torchvision
tqdm
yapf
timm
einops
kornia
matplotlib